Separate integration from unit tests
